api/requirements.txt
flask
groq
openai
cohere
python-dotenv
flask-cors
pinecone
google-genai
